kanal_servis/requirements.txt
amqp==5.1.1
APScheduler==3.6.3
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==2.2.19
django-celery-beat==2.3.0
django-timezone-field==5.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
kombu==5.2.4
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.30
protobuf==4.21.4
psycopg2==2.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
